Undo last commit and no longer warning if bounds seem invalid, for following reason:
Right now I cam calculating bounds for the smallest *local minimum* in the simplex, *not* for the smallest possible point in the simplex. It is possible to have the bound on a local minimum to be greater than the bound on the smallest value in the function. Consider a function that takes the values 0 and 1 at x=0 and x=1, and whose second derivative is bounded from above by 0.5 (i.e the Lipshitz constant for grad(f) is 0.5). By the MVT, grad(f) is 1 at at least 1 point on x = (0, 1). At a minimum, grad(f) = 0. So if there is a minimum, then by the MVT grad(grad(f)) is at least 1, which contradicts the bound on grad(grad(f)). In other words, there *cannot* be a local minimum on the interval. For a simplex, it is more complicated, but it is still possible to get a bound for a local minimum that is greater than the minimum value.